remove float32 materialization

+++ models/llama_mustafar_kernel.py
@@ -890,7 +890,7 @@ class LlamaForCausalLM_MUSTAFAR(LlamaPreTrainedModel):
             logits = torch.cat(logits, dim=-1)
         else:
             logits = self.lm_head(hidden_states)
-        logits = logits.float()
+        # logits = logits.float()
 
         loss = None
         if labels is not None:
